Fix insertion coloring, recursion and removal with two children
Some additional tests caught problems around insertion coloring and
balancing as ell as recursion and balancing. One case caused a crash,
and another caused the tree to be imbalanced. Both of these
were fixed. The removal case was completely missing the branch to
handle a node with two children. This has been added as well.